slack-bot/pyproject.toml
[project]
name = "studio-kaku-slack-bot"
version = "0.1.0"
description = "Slack bot for Studio Kaku website editing via Claude Code"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.20",
    "aiohttp>=3.9",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]
